refactor(ui): move shell rendering out of app
